% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{prevail-globals}
\alias{prevail-globals}
\title{Declare Global Variables for Package}
\description{
Declares global variables used in the package functions to avoid R CMD check notes about "no visible binding".
}
\keyword{internal}
