% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_vaccination_data.R
\name{case_vaccine_to_param}
\alias{case_vaccine_to_param}
\title{Format Case and Vaccination Parameters for Model Input (Including Pre-1980)}
\usage{
case_vaccine_to_param(
  demog_data,
  processed_vaccination,
  processed_vaccination_sia,
  processed_case,
  vaccination_schedule,
  vaccination_pre1980,
  vaccine_parameters,
  disease_parameters,
  WHO_seed_switch
)
}
\arguments{
\item{demog_data}{Output list from `process_demography()`.}

\item{processed_vaccination}{Data frame of routine vaccine coverage.}

\item{processed_vaccination_sia}{Data frame of SIA coverage.}

\item{processed_case}{WHO case reports processed for the country of interest.}

\item{vaccination_schedule}{Full WHO vaccine schedule data frame.}

\item{vaccination_pre1980}{Data frame of historical vaccination introduction assumptions.}

\item{vaccine_parameters}{Data frame of vaccination parameters}
}
\value{
A named list of parameters including coverage arrays and seeding inputs.
}
\description{
Wrapper function that prepares all time-varying parameters for use in an age-, vaccine-,
and risk-structured infectious disease model, including extrapolated vaccination coverage
prior to 1980 if available.
}
\keyword{internal}
