% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat_parameters.R
\name{reformat_parameters}
\alias{reformat_parameters}
\title{Reformat Parameter Table for Model Input}
\usage{
reformat_parameters(
  custom_parameters,
  default_parameters,
  disease,
  type = c("disease", "vaccine")
)
}
\arguments{
\item{custom_parameters}{Data frame of custom parameters.}

\item{default_parameters}{Data frame of the default parameters}

\item{disease}{Character string, name of the disease.}

\item{type}{One of \code{"disease"} or \code{"vaccine"}. Controls formatting logic.}
}
\value{
A data frame with reformatted parameters.
}
\description{
Reformats a parameter table for PREVAIL, standardising column names and units for either disease or vaccine parameters.
}
